+++ R/as_bm_pixmap.R
@@ -30,10 +30,10 @@ as_bm_pixmap.default <- function(x, ...) {
 #' @rdname as_bm_pixmap
 #' @export
 as_bm_pixmap.bm_bitmap <- function(x, ..., col = getOption("bittermelon.col", col_bitmap)) { # nolint
-    col <- col2rrggbbaa(col)
-    x <- apply(as.matrix(x), 2L, function(i) col[i + 1L])
-    class(x) <- c("bm_pixmap", "bm_matrix", class(x))
-    x
+    cols <- col2rrggbbaa(col)[as.integer(x) + 1L]
+    m <- matrix(cols, nrow = nrow(x), ncol = ncol(x))
+    class(m) <- c("bm_pixmap", "bm_matrix", class(m))
+    m
 }
 
 #' @rdname as_bm_pixmap
@@ -46,11 +46,10 @@ as_bm_pixmap.bm_pixmap <- function(x, ...) {
 #' @export
 as_bm_pixmap.matrix <- function(x, ...) {
     if (is.character(x)) {
-        if (nrow(x) > 0L && ncol(x) > 0L)
-            for (i in seq_len(nrow(x)))
-                x[i, ] <- col2rrggbbaa(x[i, ])
-        class(x) <- c("bm_pixmap", "bm_matrix", class(x))
-        x
+        cols <- col2rrggbbaa(as.character(x))
+        m <- matrix(cols, nrow = nrow(x), ncol = ncol(x))
+        class(m) <- c("bm_pixmap", "bm_matrix", class(m))
+        m
     } else {
         as_bm_pixmap.bm_bitmap(as_bm_bitmap.matrix(x))
     }
@@ -62,7 +61,6 @@ as_bm_pixmap.matrix <- function(x, ...) {
 #' @export
 as_bm_pixmap.maze <- function(x, ..., walls = FALSE, start = NULL, end = NULL,
                               col = getOption("bittermelon.col", col_bitmap)) {
-    stopifnot(requireNamespace("mazing", quietly = TRUE))
     as_bm_pixmap.bm_bitmap(as_bm_bitmap.maze(x, walls = walls, start = start, end = end),
                            col = col)
 }
@@ -78,12 +76,11 @@ as_bm_pixmap.nativeRaster <- function(x, ...) {
     stopifnot(requireNamespace("farver", quietly = TRUE))
     if (nrow(x) > 0L && ncol(x) > 0L) {
         cols <- col2rrggbbaa(farver::decode_native(as.integer(x)))
-        m <- matrix(cols, nrow = nrow(x), ncol = ncol(x), byrow = TRUE)
-        m <- m[seq.int(nrow(x), 1L), , drop = FALSE]
+        m <- flip_matrix_vertically(matrix(cols, nrow = nrow(x), ncol = ncol(x), byrow = TRUE))
         class(m) <- c("bm_pixmap", "bm_matrix", class(m))
         m
     } else {
-        as_bm_pixmap.matrix("#FFFFFF00", nrow = nrow(x), ncol = ncol(x))
+        as_bm_pixmap.matrix(matrix("#FFFFFF00", nrow = nrow(x), ncol = ncol(x)))
     }
 }
 
@@ -92,12 +89,11 @@ as_bm_pixmap.nativeRaster <- function(x, ...) {
 as_bm_pixmap.raster <- function(x, ...) {
     # Standardize all colors to #RRGGBBAA format
     if (nrow(x) > 0L && ncol(x) > 0L) {
-        x <- as.matrix(x)[seq.int(nrow(x), 1L), , drop = FALSE]
-        for (i in seq_len(nrow(x)))
-            x[i, ] <- col2rrggbbaa(x[i, ])
-        class(x) <- c("bm_pixmap", "bm_matrix", class(x))
-        x
+        cols <- col2rrggbbaa(as.character(flip_matrix_vertically(as.matrix(x))))
+        m <- matrix(cols, nrow = nrow(x), ncol = ncol(x))
+        class(m) <- c("bm_pixmap", "bm_matrix", class(m))
+        m
     } else {
-        as_bm_pixmap.matrix("#FFFFFF00", nrow = nrow(x), ncol = ncol(x))
+        as_bm_pixmap.matrix(matrix("#FFFFFF00", nrow = nrow(x), ncol = ncol(x)))
     }
 }